Pass form instance to "entity" field queryBuilder
This makes it possible to react to the current model values in
the query builder.

Example:
A "category" dropdown of type "entity".
Its queryBuilder adjusts the query to only show public categories.
When editing a model that already has a non-public category selected,
the query builder is now able to take this into account and
show public categories plus the currently selected non-public/internal
category.